# openapi.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'workorder-ri.appspot.com'
basePath: /
tags:
  - name: work-order-controller
    description: Work Order Controller
  - name: ui-controller
    description: UI Controller
paths:
  /:
    get:
      tags:
        - ui-controller
      summary: loadIndexPage
      operationId: loadIndexPageUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/changeStatus/{id}':
    post:
      tags:
        - ui-controller
      summary: processForm
      operationId: processFormUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /getWorkOrder:
    get:
      tags:
        - work-order-controller
      summary: getWorkOrder
      operationId: getWorkOrderUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/WorkOrder'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /getWorkOrders:
    get:
      tags:
        - work-order-controller
      summary: getWorkOrders
      operationId: getWorkOrdersUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/WorkOrder'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /postWorkOrders:
    post:
      tags:
        - work-order-controller
      summary: createWorkOrders
      operationId: createWorkOrdersUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: workOrderList
          description: workOrderList
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/WorkOrder'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/WorkOrder'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /updateOrder:
    post:
      tags:
        - work-order-controller
      summary: updateWorkOrders
      operationId: updateWorkOrdersUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: workOrder
          description: workOrder
          required: true
          schema:
            $ref: '#/definitions/WorkOrder'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  WorkOrder:
    type: object
    properties:
      customerid:
        type: string
      details:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
      start_date:
        type: string
      status:
        type: string